+++ csrc/eve.c
@@ -232,13 +232,17 @@ int main(int argc, char **argv)
 
     http_server h = create_http_server(create_station(0, port), server_eve);
     register(h, "/", "text/html", index);
-    register(h, "/jssrc/renderer.js", "application/javascript", renderer);
-    register(h, "/jssrc/microReact.js", "application/javascript", microReact);
-    register(h, "/jssrc/codemirror.js", "application/javascript", codemirror);
-    register(h, "/jssrc/codemirror.css", "text/css", codemirrorCss);
+    register(h, "/js/microReact.js", "application/javascript", microReact);
+    register(h, "/js/codemirror.js", "application/javascript", codemirror);
+    register(h, "/js/codemirror.css", "text/css", codemirrorCss);
     register(h, "/examples/todomvc.css", "text/css", exampleTodomvcCss);
-    register(h, "/jssrc/commonmark.js", "application/javascript", commonmark);
-    register(h, "/jssrc/editor.js", "application/javascript", editor);
+    register(h, "/js/commonmark.js", "application/javascript", commonmark);
+    register(h, "/js/system.js", "application/javascript", systemjs);
+
+    register(h, "/js/util.js", "application/javascript", client);
+    register(h, "/js/client.js", "application/javascript", client);
+    register(h, "/js/renderer.js", "application/javascript", renderer);
+    register(h, "/js/editor.js", "application/javascript", editor);
 
     // TODO: figure out a better way to manage multiple graphs
     init_json_service(h, root, enable_tracing, compiler_bag, exec_path);
