+++ jssrc/declarations.d.ts
@@ -1,0 +1,3 @@
+declare module "commonmark";
+declare module "CodeMirror";
+declare module "microReact";
