+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "module": "system",
+    "moduleResolution": "node",
+    "outDir": "build",
+    "target": "es5",
+    "inlineSourceMap": true,
+    "skipDefaultLibCheck": true,
+
+    "forceConsistentCasingInFileNames": true
+  },
+  "include": [ "jssrc/**/*.ts" ]
+}
